fix(strategies): fix neighboring regions with same owner.
Neighboring region with the same owner should be merged together because
otherwise processes that request memory space will appear to think that
there's only a portion of the memory available, not to mention allocators
aren't aware of that the space should be treated as one either.

Whenver we deallocate, merge neighboring regions.

fixes d6fb38b0